+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/feature_wrapper/global/esf_estimator.h
@@ -25,7 +25,7 @@ namespace pcl
         void
         estimate (PointInTPtr & in, PointInTPtr & processed,
                   typename pcl::PointCloud<FeatureT>::CloudVectorType & signatures,
-                  std::vector<Eigen::Vector3f> & centroids)
+                  std::vector<Eigen::Vector3d> & centroids)
         {
 
           typedef typename pcl::ESFEstimation<PointInT, FeatureT> ESFEstimation;
@@ -40,9 +40,9 @@ namespace pcl
 
           signatures[0] = ESF_signature;
 
-          Eigen::Vector4f centroid4f;
+          Eigen::Vector4d centroid4f;
           pcl::compute3DCentroid (*in, centroid4f);
-          centroids[0] = Eigen::Vector3f (centroid4f[0], centroid4f[1], centroid4f[2]);
+          centroids[0] = Eigen::Vector3d (centroid4f[0], centroid4f[1], centroid4f[2]);
 
           pcl::copyPointCloud(*in, *processed);
           //processed = in;
