+++ apps/in_hand_scanner/include/pcl/apps/in_hand_scanner/integration.h
@@ -104,7 +104,7 @@ namespace pcl
         bool
         merge (const CloudXYZRGBNormalConstPtr& cloud_data,
                MeshPtr&                         mesh_model,
-               const Eigen::Matrix4f&           T) const;
+               const Eigen::Matrix4d&           T) const;
 
         /** \brief Outlier rejection. In each merge step points that have not been observed again age by one iteration. Points that are observed again get an age of 0. Once a point reaches the maximum age it is decided if the point is removed or kept in the mesh. A point is removed if it has not been observed from a minimum number of directions.
           * \param[in,out] mesh The mesh which should be processed.
